refactor: deduplicate reference data load errors
